backend/requirements.txt
flask==3.1.3
psycopg2-binary==2.9.11
